refactor dialogs into /game/view
