+++ app/src/androidTest/java/com/github/fribourgsdp/radio/LyricsPopupTest.kt
@@ -1,5 +1,4 @@
 package com.github.fribourgsdp.radio
-import androidx.fragment.app.testing.FragmentScenario
 import androidx.fragment.app.testing.launchFragment
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions
@@ -8,8 +7,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.RootMatchers.isDialog
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.hamcrest.Matchers.not
-import org.junit.Assert.assertEquals
+import com.github.fribourgsdp.radio.game.view.LyricsPopup
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -21,7 +19,7 @@ class LyricsPopupTest {
     fun testCorrectTextIsDisplayed(){
         with(launchFragment(
             themeResId = R.style.Theme_Radio,
-            instantiate = {LyricsPopup("Lorem Ipsum")})) {
+            instantiate = { LyricsPopup("Lorem Ipsum") })) {
             onView(withId(R.id.lyricsPopupTextView))
                 .inRoot(isDialog())
                 .check(matches(withText("Lorem Ipsum")))
@@ -31,7 +29,7 @@ class LyricsPopupTest {
     fun testPopupClosingOnButtonPress(){
         with(launchFragment(
             themeResId = R.style.Theme_Radio,
-            instantiate = {LyricsPopup("Lorem Ipsum")})) {
+            instantiate = { LyricsPopup("Lorem Ipsum") })) {
             onView(withId(R.id.close_popup_button))
                 .inRoot(isDialog())
                 .perform(ViewActions.click())
