+++ app/src/androidTest/java/com/github/fribourgsdp/radio/SongPickerDialogTest.kt
@@ -9,6 +9,7 @@ import androidx.test.espresso.matcher.RootMatchers.isDialog
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.fribourgsdp.radio.game.GameView
+import com.github.fribourgsdp.radio.game.view.SongPickerDialog
 import org.hamcrest.Matchers.not
 import org.junit.Assert.assertEquals
 import org.junit.Test
