+++ app/src/main/java/com/github/fribourgsdp/radio/game/view/QuitGameOrLobbyDialog.kt
@@ -1,9 +1,6 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.game.view
 
-import android.app.Activity
 import android.content.Context
-import android.content.ContextWrapper
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -12,6 +9,7 @@ import android.widget.Button
 import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.setFragmentResult
+import com.github.fribourgsdp.radio.R
 
 
 class QuitGameOrLobbyDialog(val ctx: Context): DialogFragment() {
